+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASE]
 
+- Fix incompatibility of `multiple` dropdowns with `massiveaction`
+
 ## [1.21.22] - 2025-05-28
 
 ### Fixed